+++ wgpu-hal/deko3d-sys/abi/deko3d_sys_b1_abi_check.cpp
@@ -26,6 +26,7 @@ CHECK_FN(dkCmdBufBindShaders, void (*)(DkCmdBuf, uint32_t, DkShader const* const
 CHECK_FN(dkCmdBufBindRasterizerState, void (*)(DkCmdBuf, DkRasterizerState const*));
 CHECK_FN(dkCmdBufBindColorState, void (*)(DkCmdBuf, DkColorState const*));
 CHECK_FN(dkCmdBufBindColorWriteState, void (*)(DkCmdBuf, DkColorWriteState const*));
+CHECK_FN(dkCmdBufBindBlendStates, void (*)(DkCmdBuf, uint32_t, DkBlendState const*, uint32_t));
 CHECK_FN(dkCmdBufBindDepthStencilState, void (*)(DkCmdBuf, DkDepthStencilState const*));
 CHECK_FN(dkCmdBufPushData, void (*)(DkCmdBuf, DkGpuAddr, void const*, uint32_t));
 CHECK_FN(dkCmdBufBindTextures, void (*)(DkCmdBuf, DkStage, uint32_t, DkResHandle const*, uint32_t));